pyproject.toml
[project]
name = "bookmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "fastapi[all]>=0.128.0",
    "langchain>=1.2.6",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.6",
    "langgraph-checkpoint-postgres>=3.0.3",
    "loguru>=0.7.3",
    "psycopg[binary,pool]>=3.3.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.3",
    "supabase>=2.27.2",
    "tavily-python>=0.7.19",
    "playwright>=1.57.0",
    "playwright-stealth>=2.0.1",
    "markitdown[all]>=0.0.2",
    "torch==2.6.0",
    "numpy==1.26.4",
    "python-slugify>=8.0.4",
    "exa-py>=2.2.0",
    "notebooklm-mcp-server>=0.1.15",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.9",
    "psycopg2-binary>=2.9.11",
    "aiogram>=3.25.0",
    "aiohttp>=3.13.3",
    "aiofiles>=25.1.0",
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.8",
]
